chore: simplify components
