+++ i18n/en.pot
@@ -5,8 +5,8 @@ msgstr ""
 "Content-Type: text/plain; charset=utf-8\n"
 "Content-Transfer-Encoding: 8bit\n"
 "Plural-Forms: nplurals=2; plural=(n != 1)\n"
-"POT-Creation-Date: 2019-02-13T13:32:16.100Z\n"
-"PO-Revision-Date: 2019-02-13T13:32:16.100Z\n"
+"POT-Creation-Date: 2019-06-06T08:29:40.053Z\n"
+"PO-Revision-Date: 2019-06-06T08:29:40.053Z\n"
 
 msgid "Cancel"
 msgstr ""
@@ -17,7 +17,7 @@ msgstr ""
 msgid "Open user guide"
 msgstr ""
 
-msgid "Scheduler app"
+msgid "Scheduler"
 msgstr ""
 
 msgid "Delete"
