+++ public/index.html
@@ -6,6 +6,6 @@
     <title>DHIS2 Scheduler</title>
 </head>
 <body>
-    <div id="scheduler"></div>
+    <div id="root"></div>
 </body>
 </html>
